+++ docs/md/releasenotes.md
@@ -16,6 +16,7 @@
 - Fix inconsistent usage of type names in the context of the `PartialEvaluator` that could result in
   trying to create an internal type string from a string that was already an internal type. (`T15513`)
 - Fix initialization of Kotlin callable references when using `-Xno-optimized-callable-references` compiler option. (`T16486`)
+- Fix `createCache()` delegation in `ArgBamCpaRun`
   
 ### Upgrade considerations
 ####TYPE NAME CONVENTION